src/requirements.txt
fastapi
uvicorn
pandas
agno
pylance
lancedb
openai
duckduckgo-search
maturin
tantivy
aiofiles